fix: observations solved
