+++ src/schemas/eventSchema.ts
@@ -252,8 +252,7 @@ export const eventRegistrationsSchema = z.object({
   updatedAt: z.date(),
   paid: z.boolean(),
   paymentId: z.string().optional().nullable(),
-  customFields: z.any().optional().nullable(),
-  eventId: z.string()
+  customFields: z.any().optional().nullable()
 })
 
 export type EventRegistrationsSchemaType = z.infer<
