+++ src/services/actions/event/updateRegistrationStatus.ts
@@ -2,7 +2,29 @@
 
 import { getUser, isOrganizerAdmin } from '@/lib/auth/lucia'
 import prisma from '@/lib/prisma'
+import { AttendeeStatus } from '@prisma/client'
 import { revalidatePath } from 'next/cache'
+
+export const updateAttendeeTicketsByStatus = async (
+  eventId: string,
+  userId: string,
+  attendeeStatus: AttendeeStatus
+) => {
+  const ticketOfferings = await prisma.ticketOfferings.findMany({
+    where: { eventId },
+    select: { id: true }
+  })
+
+  await prisma.ticket.updateMany({
+    where: {
+      AND: [
+        { userId },
+        { ticketTypeId: { in: ticketOfferings.map(_ => _.id) } }
+      ]
+    },
+    data: { deletedAt: attendeeStatus === 'APPROVED' ? null : new Date() }
+  })
+}
 
 export async function updateRegistrationStatus(
   registrationId: string,
@@ -31,6 +53,12 @@ export async function updateRegistrationStatus(
       where: { id: registrationId },
       data: { status }
     })
+
+    await updateAttendeeTicketsByStatus(
+      registration.eventId,
+      registration.userId,
+      status
+    )
 
     if (status === 'APPROVED') {
       await prisma.queueJob.create({
